Redo gpuequiv benchmarks
The numbers now represent the final state of gpuequiv version 1.0.0.
The game graph shrunk a bit again, which caused some small performance
improvements.

The number of significant digits was reduced to better correlate to the
variance in benchmarking.

The runtime for generating the game graph of vasy_18_73 was removed from
the table, because the full game was never run, which was misleading.